Add ability to open IM channels by user name
When sending messages to an IM channel, clients are expected to send the channel ID. But when receiving messages from an IM channel, they only get the name of the user the IM channel is open to. This commit adds a get_im_channel method which translates a user name into the IM channel's ID, so that RTM clients can readily manage conversation over IM channels. In order to make this change, it adds a private _find_user_id method to the SlackSocket class, because message events on IM channels contain the user name, but the user ID is required to get the channel.

This commit also adds the ability to open an IM channel to a particular user. The get_im_channel method accepts a username, looks up the user ID, and checks the local channel cache to find the IM channel to that user. If none exists, it creates one, by calling the new _open_im_channel private method.

The SlackClient class has been modified to support making POST requests as well as GET requests, so that the SlackSocket can make calls to the RTM API to open IM channels. Specifically, the get_json method now accepts an HTTP method and a parameter list as arguments. Unused private methods have also been removed.

Lastly, a URL to the 'im.open' method has been added to the slackurl dictionary.